minor: split openai & codex responses.
